Create sub-namespaces for Client and Repository tests
