+++ test/LionWeb.Protocol.Delta.Test/Repository/RepositoryConnector.cs
@@ -15,13 +15,14 @@
 // SPDX-FileCopyrightText: 2024 TRUMPF Laser SE and other contributors
 // SPDX-License-Identifier: Apache-2.0
 
-namespace LionWeb.Protocol.Delta.Test;
+namespace LionWeb.Protocol.Delta.Test.Repository;
 
+using Client;
 using Core;
 using Core.Notification;
+using Delta.Repository;
 using Message;
 using Message.Event;
-using Repository;
 using System.Text;
 
 class RepositoryConnector : IDeltaRepositoryConnector
